% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_nhpp.R
\name{fit.vbsrm2}
\alias{fit.vbsrm2}
\title{Fit the model via the variational Bayes}
\usage{
fit.vbsrm2(
  time = NULL,
  fault = NULL,
  type = NULL,
  te = NULL,
  data = data.frame(),
  alpha = NA,
  prior.omega = list(shape = 0, rate = 1),
  prior.beta = list(shape = 0, rate = 1),
  control = list(),
  ...
)
}
\description{
Obtain the posterior distribution
}
