% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vbsrm}
\alias{vbsrm}
\alias{vbsrm-package}
\title{vbsrm: Variational Bayes for Software Reliability Models}
\description{
This package provides estimation programs for software
reliability growth models with variational Bayes.
}
